+++ site/jaws_site/runs.py
@@ -818,12 +818,18 @@ class Run:
             logger.debug(f"Run {self.data.id}: Writing {outfiles_file}")
             write_json_file(outfiles_file, outfiles)
 
-        # update and write task log
+        # update task log
         try:
             task_log = self.task_log()
-            # this copies desired fields from the cromwell metadata into the task-log table
             task_summary_dict = metadata.task_summary_dict()
             task_log.add_metadata(task_summary_dict)
+        except Exception as error:
+            logger.error(f"Run {self.data.id}: Failed to update task-log: {error}")
+            self.update_run_status("failed", "Failed to update task-log")
+            return
+
+        # write task log
+        try:
             task_log_table = task_log.table()
         except Exception as error:
             logger.error(f"Run {self.data.id}: Failed to generate task-log: {error}")
@@ -991,6 +997,7 @@ def send_run_status_logs(session, central_rpc_client) -> None:
             data["workflow_name"] = run.data.workflow_name
         elif log.status_to == "complete":
             data["output_manifest"] = run.output_manifest()
+            data["cpu_hours"] = run.data.cpu_hours
         try:
             response = central_rpc_client.request("update_run_log", data)
         except Exception as error: